Changed BitWriter padding_bits field from u8 to usize

+++ src/io.rs
@@ -27,7 +27,7 @@ struct BitWriter {
     buffer: u8,
     len: u8,
     output: Vec<u8>,
-    padding_bits: u8,
+    padding_bits: usize,
 }
 
 impl BitWriter {
